+++ android v4.2/app/src/main/java/com/dicoding/picodiploma/libas/Variables.kt
@@ -5,9 +5,9 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Variables(
-    var temp: Int? = 0,
+    var temp: Double? = 0.00,
     var humid: Int? = 0,
     var rain: Int? = 0,
-    var wind: Int? = 0
+    var wind: Double? = 0.00
 ): Parcelable {
 }
